refactor: move getcapingeststatus into catalog utils to decouple api layer from ui (#1312)
