+++ __tests__/cap-ingest-status.utils.test.ts
@@ -5,7 +5,7 @@ import {
   HCAAtlasTrackerSourceDataset,
   REPROCESSED_STATUS,
 } from "../app/apis/catalog/hca-atlas-tracker/common/entities";
-import { getCapIngestStatus } from "../app/components/Table/components/TableCell/components/CAPIngestStatusCell/utils";
+import { getCapIngestStatus } from "../app/apis/catalog/hca-atlas-tracker/common/utils";
 
 describe("getCapIngestStatus", () => {
   it("returns NOT_REQUIRED for reprocessed source datasets", () => {
